Split build graph dependency ordering
